Change to Beta Module
Update the cmdlets to use the beta modules instead of the 1.0 modules since this was changed in a resent release.
Updated the requires statement to hard code the version of the modules required instead of running with whatever is installed.
This will eliminate issues with the script not working as expected with newer versions.