Add Permissions Requirements
Add a new page describing pre-reqs for the app.
Add a new section to all plugins, to describe what permissions are needed and how they are used.
Add an intro to the permission sections in the plugins landing page.
Add links in the various pages so that the prerequisites are more discoverable.