Update Permissions For Settings Engine
Update the list of permission needed and specify which are auto granted and which need to be granted ahead of time.
Update the auto granting PowerShell script command to the minimum set of permissions instead of all permissions since the settings engine handles itself after the core permissions are met.